Add IIS7 DeleteSite operation
