+++ product/dropkick/dropkick.csproj
@@ -99,6 +99,7 @@
     <Compile Include="Configuration\Dsl\Files\UnzipOptions.cs" />
     <Compile Include="Configuration\Dsl\Iis\IisApplicationPoolOperationProtoTask.cs" />
     <Compile Include="Configuration\Dsl\Iis\IisOperations.cs" />
+    <Compile Include="Configuration\Dsl\Iis\IisSiteOperationProtoTask.cs" />
     <Compile Include="Configuration\Dsl\Iis\ProcessModelIdentity.cs" />
     <Compile Include="Configuration\Dsl\Msmq\MsmqQueueOptions.cs" />
     <Compile Include="Configuration\Dsl\MsSrss\Extensions.cs" />
@@ -150,7 +151,8 @@
     <Compile Include="Tasks\BaseTask.cs" />
     <Compile Include="Tasks\Files\UnzipArchiveTask.cs" />
     <Compile Include="Tasks\Iis\Iis6Path.cs" />
-    <Compile Include="Tasks\Iis\Iis7OperationTask.cs" />
+    <Compile Include="Tasks\Iis\Iis7ApplicationPoolOperationTask.cs" />
+    <Compile Include="Tasks\Iis\Iis7SiteOperationTask.cs" />
     <Compile Include="Tasks\Iis\IisUtility.cs" />
     <Compile Include="Tasks\Iis\ManagedRuntimeVersion.cs" />
     <Compile Include="Tasks\CommandLine\RemoteDropkickExecutionTask.cs" />
